api/requirements.txt
amqp==5.1.1
asgiref==3.4.1
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.14
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-debug-toolbar==3.5.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.9.0
django-haystack==3.2.1
django-nine==0.2.5
django-timezone-field==5.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
elastic-transport==8.1.2
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
Faker==8.10.1
flower==1.1.0
humanize==4.2.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.2
kombu==5.2.4
MarkupSafe==2.1.1
pika==1.3.0
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tornado==6.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
